fixed bugs: 1) attack feels more reflexive now and responsive to when the animation is actually going 2) stamina regenerates correctly

+++ Platformer/Source/Platformer/PaperPlatformerCharacter.h
@@ -75,11 +75,11 @@ class PLATFORMER_API APaperPlatformerCharacter : public APaperCharacter
     
     // CURRENT attack power of the Hero
     UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = Attack)
-    int32 AttackPower;
+    float AttackPower;
     
     // BASE (without modifier) attack power of the Hero
     UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = Attack)
-    int32 BaseAttackPower;
+    float BaseAttackPower;
     
     // DURRATION of an attack power buff from an AttackPickup
     UPROPERTY(VisibleAnywhere, BluePrintReadWrite, Category = Attack)
